requirements.txt
# Core
faster-whisper==1.0.3
numpy>=1.23
sounddevice>=0.4.6
webrtcvad-wheels>=2.0.10
# For saving WAV
scipy>=1.10
# Optional: OpenAI LLM post-processing
openai>=1.40.0
# Additional utilities
pydub>=0.25.1
python-dotenv>=1.0.0
# Web interface
flask>=2.3.0
flask-cors>=4.0.0
